fix: support job snapshots + set the scope in init
Fixes various issues with unnecessary `init`'s

+++ services/orchest-webserver/client/src/api/files/useFileApi.ts
@@ -1,4 +1,3 @@
-import { defineStoreScope } from "@/store/scoped";
 import { FileRoot, fileRoots, UnpackedMove } from "@/utils/file";
 import {
   addToFileMap,
@@ -11,15 +10,15 @@ import {
 } from "@/utils/file-map";
 import { directoryLevel, dirname, isDirectory } from "@/utils/path";
 import { memoizeFor, MemoizePending } from "@/utils/promise";
-import { prune } from "@/utils/record";
 import { hasValue } from "@orchest/lib-utils";
+import create from "zustand";
 import { filesApi, TreeNode } from "./fileApi";
 
 export type FetchNodeParams = {
   root: FileRoot;
   path: string | undefined;
   depth: number;
-  overrides?: FileApiOverrides;
+  overrides?: FileScope;
 };
 
 type FileRoots = Partial<Record<FileRoot, FileMap>>;
@@ -28,6 +27,8 @@ type FileRootEntry = readonly [FileRoot, FileMap];
 export type FileApi = {
   /** The currently available file maps, organized by root name. */
   roots: FileRoots;
+  /** Defines which project and pipeline, job, run or snapshot the files are from. */
+  scope: FileScope;
   /**
    * Fetches the provided directory and merges its entries into its corresponding root.
    * If a file path is provided, its parent directory is fetched instead.
@@ -36,19 +37,18 @@ export type FileApi = {
     (
       root: FileRoot,
       directory?: string | undefined,
-      overrides?: FileApiOverrides
+      overrides?: FileScope
     ) => Promise<void>
   >;
+  /** Reloads the available file roots up to their maximum expanded depth. */
+  refresh: MemoizePending<() => Promise<void>>;
   /**
    * Fetches (or reloads) all file roots up to a certain depth.
    * Anything deeper than the provided depth will be pruned in the updated roots.
    * @param depth How many levels of subdirectories to include, if omitted (or undefined) the current maximum depth of the root is used.
    */
   init: MemoizePending<
-    (
-      depth?: number,
-      overrides?: FileApiOverrides
-    ) => Promise<Record<string, FileMap>>
+    (depth: number, scope: FileScope) => Promise<Record<string, FileMap>>
   >;
   delete: MemoizePending<(root: FileRoot, path: string) => Promise<void>>;
   move: MemoizePending<(move: UnpackedMove) => Promise<void>>;
@@ -61,42 +61,25 @@ export type FileApi = {
   create: MemoizePending<(root: FileRoot, path: string) => Promise<void>>;
 };
 
-const additionalScope = ["jobUuid", "runUuid", "pipelineUuid"] as const;
-
-const create = defineStoreScope({
-  requires: ["projectUuid"],
-  additional: additionalScope,
-});
-
-export type FileApiOverrides = {
+export type FileScope = {
+  projectUuid?: string;
   jobUuid?: string;
   runUuid?: string;
   pipelineUuid?: string;
+  snapshotUuid?: string;
 };
 
 export const useFileApi = create<FileApi>((set, get) => {
-  const fetchNode = async ({
+  const fetchNode = ({
     root,
     path,
     depth,
-    overrides,
-  }: FetchNodeParams) => {
-    const { projectUuid, pipelineUuid, jobUuid, runUuid } = prune({
-      ...get(),
-      ...overrides,
-    });
+  }: FetchNodeParams): Promise<TreeNode | undefined> => {
+    const { projectUuid, ...scope } = get().scope;
 
-    if (!projectUuid) return undefined;
+    if (!projectUuid) return Promise.resolve(undefined);
 
-    return await filesApi.fetchNode({
-      root,
-      path,
-      depth,
-      projectUuid,
-      pipelineUuid,
-      jobUuid,
-      runUuid,
-    });
+    return filesApi.fetchNode({ ...scope, projectUuid, root, path, depth });
   };
 
   const getDepth = (root: string) =>
@@ -117,15 +100,12 @@ export const useFileApi = create<FileApi>((set, get) => {
 
   return {
     roots: {},
-    expand: memoizeFor(500, async (root, directory = "/", overrides) => {
+    scope: {},
+    expand: memoizeFor(500, async (root, directory = "/") => {
       directory = isDirectory(directory) ? directory : dirname(directory);
 
-      const node = await fetchNode({
-        root,
-        path: directory,
-        depth: 1,
-        overrides,
-      });
+      const node = await fetchNode({ root, path: directory, depth: 1 });
+
       if (!node) return;
 
       const contents = Object.keys(createFileMap(node)).filter(
@@ -137,7 +117,9 @@ export const useFileApi = create<FileApi>((set, get) => {
       );
     }),
     create: memoizeFor(500, async (root, path) => {
-      const { projectUuid } = get();
+      const { projectUuid } = get().scope;
+
+      if (!projectUuid) return;
 
       if (isDirectory(path)) {
         await filesApi.createDirectory({ root, path, projectUuid });
@@ -148,29 +130,29 @@ export const useFileApi = create<FileApi>((set, get) => {
       updateRoot(root, (fileMap) => addToFileMap(fileMap, path));
     }),
     delete: memoizeFor(500, async (root, path) => {
-      const { projectUuid } = get();
+      const { projectUuid } = get().scope;
+
+      if (!projectUuid) return;
 
       await filesApi.deleteNode({ projectUuid, root, path });
 
       updateRoot(root, (fileMap) => removeFromFileMap(fileMap, path));
     }),
     move: memoizeFor(500, async (move) => {
-      const { projectUuid } = get();
+      const { projectUuid } = get().scope;
+
       if (!projectUuid) return;
 
       await filesApi.moveNode(projectUuid, move);
 
       set(({ roots }) => ({ roots: moveBetween(roots, move) }));
     }),
-    init: memoizeFor(500, async (depth, overrides) => {
+    init: memoizeFor(500, async (depth, scope) => {
+      set({ scope });
+
       const entries = await Promise.all(
         fileRoots.map((root) =>
-          fetchNode({
-            root,
-            path: "/",
-            depth: depth ?? getDepth(root) ?? 2,
-            overrides,
-          })
+          fetchNode({ root, path: "/", depth })
             .then((node) => (!node ? undefined : ([root, node] as const)))
             .catch(() => undefined)
         )
@@ -183,6 +165,21 @@ export const useFileApi = create<FileApi>((set, get) => {
       set({ roots });
 
       return roots;
+    }),
+    refresh: memoizeFor(500, async () => {
+      const entries = await Promise.all(
+        fileRoots.map((root) =>
+          fetchNode({ root, path: "/", depth: getDepth(root) ?? 2 })
+            .then((node) => (!node ? undefined : ([root, node] as const)))
+            .catch(() => undefined)
+        )
+      ).then((roots) => roots.filter(hasValue));
+
+      const roots = Object.fromEntries(
+        entries.map(([root, node]) => createRootEntry(root, node))
+      );
+
+      set({ roots });
     }),
   };
 });
